Rename soroban-spec::codegen to ::gen
