+++ soroban-spec/src/gen.rs
