+++ soroban-spec/src/gen/rust.rs
