+++ soroban-spec/src/gen/rust/trait.rs
