+++ soroban-spec/src/gen/rust/types.rs
